glue_kinesis_to_hudi/requirements.txt
boto3==1.26.67
botocore==1.29.67
Faker==16.7.0
jmespath==1.0.1
pip==22.3.1
python-dateutil==2.8.2
python-dotenv==0.21.1
s3transfer==0.6.0
setuptools==65.7.0
six==1.16.0
urllib3==1.26.14
wheel==0.38.4
